Move SqliteEngine to Microsoft.Data.Sqlite.Internal
Resolves #262
